refactor(react-headless-components-preview): split usePopover and usePopoverAuto
